test(framework): make three e2e suites compile and start again
ResilienceLifecycle, SpotService and ToActorMessaging never reached a
scenario because their host projects no longer built against v11. The suite
runners send build output to /dev/null, so each failed with no message at
all. SendHighWaterMark is ulong now, Contracts.Eventing is gone,
EnableActorDispatch takes no mesh name, and ActorRef.Generation is
ObjectGeneration.

ResilienceLifecycle's provider then hit the same defect as StoreFailure: it
pinned a fixed routing ID while discovering peers through the Location Store,
which spec 13 section 3.3 allows only in an explicit manual topology. Its
consumer already used a prefix, so only the provider and the probe's exact
rid comparison needed changing.

ResilienceLifecycle hosts now start and RL-A1 reaches its assertion.
ToActorMessaging has the same fixed-RID defect but its caller connects to
actor peers by routing ID, so it needs a topology decision rather than the
same edit.

+++ framework/languages/dotnet/e2e/SpotService/Server/Gateway/GatewayHostFactory.cs
@@ -547,7 +547,7 @@ internal sealed record GatewayOptions(
     string? ExternalSpotEndpoint = null,
     string? SpotPeerAEndpoint = null,
     string? SpotPeerBEndpoint = null,
-    int SpotPublisherSendHighWaterMark = 1,
+    ulong SpotPublisherSendHighWaterMark = 1,
     int SpotPublisherSendTimeoutMilliseconds = 250)
 {
     public static GatewayOptions Parse(string[] args)
